+++ src/Utils/ObjectTreerinator/ArrayRenderer.tsx
@@ -1,6 +1,7 @@
 import React, { FunctionComponent } from 'react'
 import { Collapsible } from '@/Utils/ObjectTreerinator/Collapsible'
 import { ArrayNodeRenderer } from '@/Utils/ObjectTreerinator/ArrayNodeRenderer'
+import { TreeMatch } from './TreeMatch'
 
 interface Props {
   property: string
@@ -14,13 +15,17 @@ export const ArrayRenderer: FunctionComponent<Props> = ({
   level,
 }) => (
   <li key={property}>
-    <span role='collapsible-property'>{property}</span>
+    <span role='collapsible-property'>
+      <TreeMatch text={property} />
+    </span>
 
     <Collapsible object={child} level={level + 1}>
       <ol start={0} role='array'>
         {child.map((item, index) => (
           <li key={index} role='item'>
-            <span role='index'>{index}:</span>
+            <span role='index'>
+              <TreeMatch text={String(index)} />:
+            </span>
             {ArrayNodeRenderer(item, level + 1)}
           </li>
         ))}