+++ src/Utils/ObjectTreerinator/StringRenderer.tsx
@@ -1,8 +1,16 @@
 import React from 'react'
+import { TreeMatch } from './TreeMatch'
 
 export const StringRenderer = (key: string, child: string) => (
   <li key={key}>
-    <span role='property'>{key}</span>:&nbsp;
-    <span role='string'>{`"${child}"`}</span>
+    <span role='property'>
+      <TreeMatch text={key} />
+    </span>
+    :&nbsp;
+    <span role='string'>
+      &quot;
+      <TreeMatch text={child} />
+      &quot;
+    </span>
   </li>
 )